edit code style
