+++ my-app/src/components/content/content.css
@@ -1,5 +1,6 @@
 .content {
     width: 75%;
+    margin: 0 auto;
 }
 .content .wrapper-product {
     display: flex;
