+++ my-app/src/components/content/content.jsx
@@ -1,14 +1,10 @@
-import React, {Component} from 'react';
+import React from 'react';
 import './content.css';
 import data from '../constants.js';
 
-// why class?
-// remake into stateless component
-class Content extends Component {
-    render() {
-        // create new from it
+
+const Content = () => {
         const elem = data.content.map((item) => {
-            // make it shorter such in App.tsx
             return (
                 <div className="product-item" key={item.id}>
                     <div className="img-main-photo">
@@ -21,18 +17,15 @@ class Content extends Component {
                 </div>
             );
         });
-
         return (
-            <div className="content">
-                {/* first must be container and after content*/}
-                <div className="container">
+            <div className="container">
+                <div className="content">
                     <div className="wrapper-product">
                         {elem}
                     </div>
                 </div>
             </div>
         )
-    }
-}
+};
 
 export default Content;
