+++ my-app/src/components/mainBanner/mainBanner.css
@@ -1,19 +1,17 @@
-/* why do you use .app selector? */
-.app .mainBanner {
+.mainBanner {
     justify-content: flex-end;
     background: #f3f3f3;
     display: flex;
     align-items: center;
     padding: 0 14% 0 0;
 }
-.app .mainBanner .description {
+.mainBanner .description {
     display: flex;
     flex-direction: column;
     padding: 0 24% 0;
     align-items: center;
 }
-.app .mainBanner .description p {
-    font-family: Proxima Nova; /*missing generic font*/
+.mainBanner .description p {
     font-size: 22px;
     line-height: 40px;
     color: #030303;
