+++ my-app/src/index.css
@@ -17,4 +17,5 @@ h1 {
     margin: 0 auto;
     display: flex;
     justify-content: space-between;
+    font-family: Proxima Nova;
 }
